refactor: re-export 제거, 폴더구조 변경
- @types 대신 types사용
- index.ts내부 파일 및 코드 각 도메인 별로 이동
- re-export 제거 (에디터의 지원, 관리 포인트를 늘릴 필요 없음)
