+++ src/Routes.tsx
@@ -1,16 +1,14 @@
 import { Route, RouteProps, Redirect } from "react-router";
 
-import {
-  Login,
-  Logout,
-  Signup,
-  StationManagement,
-  LineManagement,
-  SectionManagement,
-  SubwayMap,
-} from "./pages";
+import LineManagementPage from "./pages/LineManagement/LineManagementPage";
+import LoginPage from "./pages/Login/LoginPage";
+import LogoutPage from "./pages/Logout/LogoutPage";
+import SectionManagementPage from "./pages/SectionManagement/SectionManagementPage";
+import SignupPage from "./pages/Signup/SignupPage";
+import StationManagementPage from "./pages/StationManagement/StationManagementPage";
+import SubwayMapPage from "./pages/SubwayMap/SubwayMapPage";
 
-import { PAGE_PATH } from "./constants";
+import { PAGE_PATH } from "./constants/route";
 import { RouteShape } from "./types/route";
 
 interface Props {
@@ -21,37 +19,37 @@ const ROUTES: RouteShape[] = [
   {
     isPrivate: false,
     path: PAGE_PATH.LOGIN,
-    Component: Login,
+    Component: LoginPage,
   },
   {
     isPrivate: true,
     path: PAGE_PATH.LOGOUT,
-    Component: Logout,
+    Component: LogoutPage,
   },
   {
     isPrivate: false,
     path: PAGE_PATH.SIGN_UP,
-    Component: Signup,
+    Component: SignupPage,
   },
   {
     isPrivate: true,
     path: [PAGE_PATH.HOME, PAGE_PATH.STATION_MANAGEMENT],
-    Component: StationManagement,
+    Component: StationManagementPage,
   },
   {
     isPrivate: true,
     path: PAGE_PATH.LINE_MANAGEMENT,
-    Component: LineManagement,
+    Component: LineManagementPage,
   },
   {
     isPrivate: true,
     path: PAGE_PATH.SECTION_MANAGEMENT,
-    Component: SectionManagement,
+    Component: SectionManagementPage,
   },
   {
     isPrivate: true,
     path: PAGE_PATH.SUBWAY_MANAGEMENT,
-    Component: SubwayMap,
+    Component: SubwayMapPage,
   },
 ];
