+++ src/components/Confirm/Confirm.styles.ts
@@ -1,6 +1,6 @@
 import styled from "styled-components";
 
-import { COLOR } from "../../constants";
+import { COLOR } from "../../constants/color";
 
 const Block = styled.div`
   width: 300px;
