+++ src/components/ErrorMessage/ErrorMessage.style.ts
@@ -1,8 +1,8 @@
 import styled from "styled-components";
 
-import { FlexCenter } from "../Layout/";
+import { FlexCenter } from "../Layout/FlexContainer/FlexContainer";
 
-import { COLOR } from "../../constants";
+import { COLOR } from "../../constants/color";
 
 const ErrorMessageBlock = styled(FlexCenter)`
   color: ${COLOR.RED_500};
