--- src/components/Layout/index.ts
@@ -1,17 +1,0 @@
-import {
-  Flex,
-  FlexJustfiyCenter,
-  FlexBetween,
-  FlexAlignCenter,
-  FlexCenter,
-} from "./FlexContainer/FlexContainer";
-import { Grid } from "./GridContainer/GridContainer";
-
-export {
-  Flex,
-  FlexJustfiyCenter,
-  FlexBetween,
-  FlexAlignCenter,
-  FlexCenter,
-  Grid,
-};
