+++ src/components/ListItem/ListItem.styles.ts
@@ -1,6 +1,6 @@
 import styled from "styled-components";
 
-import { COLOR } from "../../constants";
+import { COLOR } from "../../constants/color";
 
 interface CircleProps {
   circleColor?: string;
