+++ src/components/ListItem/ListItem.tsx
@@ -1,7 +1,8 @@
 import { LiHTMLAttributes } from "react";
 
-import { Button, Confirm } from "../";
-import { FlexAlignCenter } from "../Layout/";
+import { FlexAlignCenter } from "../Layout/FlexContainer/FlexContainer";
+import Button from "../Button/Button";
+import Confirm from "../Confirm/Confirm";
 
 import { useModal } from "../../hooks";
