--- src/components/index.ts
@@ -1,57 +1,0 @@
-import Block from "./Block/Block";
-import Button from "./Button/Button";
-import ColorPicker from "./ColorPicker/ColorPicker";
-import Confirm from "./Confirm/Confirm";
-import ErrorMessage from "./ErrorMessage/ErrorMessage";
-import Header from "./Header/Header";
-import Input from "./Input/Input";
-import InputField from "./Input/InputField";
-import Select from "./Select/Select";
-import SelectField from "./Select/SelectField";
-import LabeledNode from "./LabeledNode/LabeledNode";
-import ListItem from "./ListItem/ListItem";
-import Loading from "./Loading/Loading";
-import ModalProvider from "./Modal/ModalProvider";
-import FormProvider from "./Form/FormProvider";
-import Navigation from "./Navigation/Navigation";
-import Line from "./Line/Line";
-import Form from "./Form/Form";
-
-import {
-  Flex,
-  FlexJustfiyCenter,
-  FlexBetween,
-  FlexAlignCenter,
-  FlexCenter,
-  Grid,
-} from "./Layout";
-
-export {
-  Block,
-  Button,
-  ColorPicker,
-  Confirm,
-  ErrorMessage,
-  Header,
-  Input,
-  InputField,
-  Select,
-  SelectField,
-  LabeledNode,
-  ListItem,
-  Loading,
-  ModalProvider,
-  FormProvider,
-  Navigation,
-  Line,
-  Form,
-};
-
-export {
-  Flex,
-  FlexJustfiyCenter,
-  FlexBetween,
-  FlexAlignCenter,
-  FlexCenter,
-  Grid,
-};
