--- src/constants/index.ts
@@ -1,9 +1,0 @@
-import { COLOR, CIRCLE_COLOR } from "./color";
-import { PAGE_PATH, NAV_LINKS } from "./route";
-import { SIZE, Z_INDEX } from "./size";
-import { ERROR_DURATION } from "./time";
-
-export { COLOR, CIRCLE_COLOR };
-export { PAGE_PATH, NAV_LINKS };
-export { SIZE, Z_INDEX };
-export { ERROR_DURATION };
