+++ src/modules/line/thunks.ts
@@ -1,7 +1,7 @@
 import { createAction, createAsyncThunk } from "@reduxjs/toolkit";
 import { LineAddRequestItem, SectionAddRequestItem } from "../../types/line";
 import { lines, sections } from "../../apis";
-import { ERROR_DURATION } from "../../constants";
+import { ERROR_DURATION } from "../../constants/time";
 
 const resetError = createAction("[LINE] RESET_ERROR");
