+++ src/pages/LineManagement/Modal/LineManagementPage.styles.ts
@@ -1,8 +1,7 @@
 import styled from "styled-components";
 
-import { FlexCenter } from "../../../components";
-
-import { SIZE } from "../../../constants";
+import { FlexCenter } from "../../../components/Layout/FlexContainer/FlexContainer";
+import { SIZE } from "../../../constants/size";
 
 export const LineManagementPageBlock = styled(FlexCenter)`
   max-width: ${SIZE.PAGE_CONTAINER_WIDTH};
