+++ src/pages/Logout/LogoutPage.tsx
@@ -2,7 +2,7 @@ import { Redirect } from "react-router";
 
 import { useAuth } from "../../hooks";
 
-import { PAGE_PATH } from "../../constants";
+import { PAGE_PATH } from "../../constants/route";
 
 const LogoutPage = () => {
   const { logout } = useAuth();
