--- src/pages/index.ts
@@ -1,12 +1,0 @@
-import Login from "./Login/LoginPage";
-import Logout from "./Logout/LogoutPage";
-import Signup from "./Signup/SignupPage";
-
-import StationManagement from "./StationManagement/StationManagementPage";
-import LineManagement from "./LineManagement/LineManagementPage";
-import SectionManagement from "./SectionManagement/SectionManagementPage";
-import SubwayMap from "./SubwayMap/SubwayMapPage";
-
-export { Login, Logout, Signup };
-
-export { StationManagement, LineManagement, SectionManagement, SubwayMap };
