--- src/validations/index.ts
@@ -1,16 +1,0 @@
-import { validateAge } from "./age";
-import { validateEmail } from "./email";
-import { validatePassword, validatePasswordConfirm } from "./password";
-
-import { validateStationName } from "./station";
-import { validateLineName } from "./line";
-import { validateSectionDistance } from "./section";
-
-export {
-  validateAge,
-  validateEmail,
-  validatePassword,
-  validatePasswordConfirm,
-};
-
-export { validateStationName, validateLineName, validateSectionDistance };
